chore: Add unit tests
